Remove kafka integration
