--- Dockerfiles/Dockerfile.consumer
@@ -1,11 +1,0 @@
-FROM python:3.9
-
-WORKDIR /app
-
-# copy project
-COPY . .
-RUN pip install .
-
-RUN FLIT_ROOT_INSTALL=1 make install
-
-ENTRYPOINT ["udata-search-service", "consume-kafka"]
