+++ pyproject.toml
@@ -17,13 +17,11 @@ dependencies = [
     "Flask==2.1.1",
     "flit==3.6.0",
     "gunicorn==20.1.0",
-    "kafka-python==2.0.2",
     "pydantic==1.9.0",
     "pytest==6.2.5",
     "pytest-flask==1.2.0",
     "markdown==3.3.3",
     "beautifulsoup4==4.10.0",
-    "udata_event_service==0.0.8"
 ]
 
 [project.urls]
