+++ udata_search_service/infrastructure/services.py
@@ -35,6 +35,9 @@ class OrganizationService:
         except TypeError:
             return None
 
+    def delete_one(self, organization_id: str) -> None:
+        self.search_client.delete_one_organization(organization_id)
+
     @staticmethod
     def format_filters(filters):
         filtered = {k: v for k, v in filters.items() if v is not None}
@@ -79,6 +82,9 @@ class DatasetService:
             return Dataset.load_from_dict(self.search_client.find_one_dataset(dataset_id))
         except TypeError:
             return None
+
+    def delete_one(self, dataset_id: str) -> None:
+        self.search_client.delete_one_dataset(dataset_id)
 
     @staticmethod
     def format_filters(filters):
@@ -134,6 +140,9 @@ class ReuseService:
             return Reuse.load_from_dict(self.search_client.find_one_reuse(reuse_id))
         except TypeError:
             return None
+
+    def delete_one(self, reuse_id: str) -> None:
+        self.search_client.delete_one_reuse(reuse_id)
 
     @staticmethod
     def format_filters(filters):
